requirements.txt
antlr4-python3-runtime==4.12.0
antlr4-tools==0.2
astroid==2.15.4
black==23.3.0
click==8.1.3
dill==0.3.6
install-jdk==1.0.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pylint==2.17.4
pyright==1.1.306
tomlkit==0.11.8
typing_extensions==4.5.0
wrapt==1.15.0
